+++ test/runtests.jl
@@ -2910,5 +2910,19 @@ else  # run HiGHS tests
             @test length(r["Messages"]["errors"]) > 0
             @test length(r["Messages"]["warnings"]) > 0
         end
+
+        @testset "Storage Duration" begin
+            ## Battery storage
+            d = JSON.parsefile("scenarios/pv_storage.json")
+            d["ElectricStorage"]["min_duration_hours"] = 8
+            d["ElectricStorage"]["max_duration_hours"] = 8
+            s = Scenario(d)
+            inputs = REoptInputs(s)
+            m = Model(optimizer_with_attributes(HiGHS.Optimizer, "output_flag" => false, "log_to_console" => false))
+            r = run_reopt(m, inputs)
+            # Test battery size_kwh = size_hw * duration
+            @test r["ElectricStorage"]["size_kw"]*8 - r["ElectricStorage"]["size_kwh"] ≈ 0.0 atol = 0.1
+
+        end
     end
 end